cStringIO.tell needs to return an int for it to be used by __len__ (perhaps that should be relaxed instead?). Found by testing against PyAMF which uses cStringIO as a backing store for its BufferedByteStream objs.

+++ src/org/python/modules/cStringIO.java
@@ -160,7 +160,7 @@ public class cStringIO {
          * Return the file position.
          * @returns     the position in the file.
          */
-        public long tell() {
+        public int tell() {
             _complain_ifclosed();
             return pos;
         }
